Refactor sentiment analysis into multiple modules
- Increase readability
- Make functionalities more flexible and modular
- Use Pydantic model to simplify handling results
- Improve handling translation API rate limit
- Improve performance by filtering on stock change before detecting
language
- Explicitly add Swedish prefix to the vader sentiment analyser

The main function was getting a bit large, clunky and hard to follow.
Pydantic models let us take advantage of automatic validation and type
conversion.
Both `vaderSentiment` and `vaderSentiment-swedish` are imported as
`vaderSentiment`, you don't seem to be able to easily use both at the
same time. Therefor it seems important to explicitly clarify that the
application uses the Swedish package.